WIP moved blockState-modifying methods to BlockEditor
