+++ packages/reports/src/reports/blocks/Block.jsx
@@ -6,15 +6,7 @@ import {HiOutlineCode} from "react-icons/hi";
 
 /* components */
 import BlockEditor from "./BlockEditor";
-import RichTextEditor from "../editors/RichTextEditor";
-import AceWrapper from "../editors/AceWrapper";
-import ApiInput from "../components/ApiInput";
-import BlockPreview from "./BlockPreview";
-import BlockSettings from "./BlockSettings";
-import BlockOutputPanel from "./BlockOutputPanel";
-import SimpleUI from "../editors/SimpleUI";
 import ConsumerMenu from "../components/ConsumerMenu";
-import VariableList from "./VariableList";
 
 /* utils */
 import deepClone from "../../utils/js/deepClone";
@@ -26,7 +18,7 @@ import {useVariables} from "../hooks/blocks/useVariables";
 import {updateEntity} from "../../actions/reports";
 
 /* enums */
-import {ENTITY_TYPES, BLOCK_MAP, BLOCK_TYPES, BLOCK_FIELDS} from "../../utils/consts/cms";
+import {ENTITY_TYPES, BLOCK_FIELDS, BLOCK_LOGIC_TYPES} from "../../utils/consts/cms";
 import {REQUEST_STATUS} from "../../utils/consts/redux";
 
 const MODES = {
@@ -35,30 +27,27 @@ const MODES = {
 };
 
 /**
- * Most blocks have translatable content, and store their locale-specific copies in a content table.
- * Generators and vizes are the exception - they store their one & only version directly in the psql block.
- */
-const hasNoLocaleContent = type => [BLOCK_TYPES.GENERATOR, BLOCK_TYPES.VIZ].includes(type);
-
-/**
  * A Block is a visual element of any kind embedded in a Section. It can be a stat, generator,
- * selector, or anything listed in BLOCK_TYPES.
- * id - the id for this block
+ * selector, or anything listed in BLOCK_TYPES
  */
 function Block({id, modified, callbacks, inline}) {
 
+  const dispatch = useDispatch();
   const {setOpened, setModified, maybeCloseWithoutSaving, setInlineId} = callbacks;
 
-  const dispatch = useDispatch();
 
-  /* redux */
+  /* REDUX SELECTORS */
+
+  /** Locale key that has been set as default */
   const localeDefault = useSelector(state => state.cms.status.localeDefault);
-  const blocks = useSelector(state => state.cms.reports.entities.blocks);
 
   /** Block data object from database */
-  const block = blocks[id];
-  // todo1.0 fix formatters, maybe move this
-  const formatters = useSelector(state => state.cms.formatters);
+  const block = useSelector(state => state.cms.reports.entities.blocks)?.[id];
+
+  // deep clone block state on mount
+  useEffect(() => setBlockState(deepClone(block)), []);
+
+  /* LOCAL STATE VARIABLES */
 
   /**
    * The content of the entire CMS is kept in a normalized redux object called reports.
@@ -68,13 +57,48 @@ function Block({id, modified, callbacks, inline}) {
    * be persisted to psql, pulled back into redux, and redistributed as props again.
    */
   const [blockState, setBlockState] = useState(false);
+
+  // block execution if there is no block state
+  if (!block || !blockState) return null;
+
+  /**
+   * Variable that indicates where a block's "content" is stored
+   *
+   * Most blocks have translatable content, and store their locale-specific copies in a content table.
+   * Generators and vizes are the exception - they store their one & only version directly in the psql block.
+   */
+  const blockHasNoLocaleContent = Object.values(BLOCK_LOGIC_TYPES).includes(block.type);
+
   const [loading, setLoading] = useState(false);
 
+  // get input variables that this block is subscribed to
   const {variables} = useVariables(id);
 
-  useEffect(() => setBlockState(deepClone(block)), []);
+  /** This blockState's content object */
+  const blockContent = blockHasNoLocaleContent ? blockState?.content : blockState.contentByLocale[localeDefault]?.content;
 
-  if (!block || !blockState) return null;
+  /**
+   * Generic method for updating the content of the current block's working state.
+   * Meant to be used by the various editors that a Block might have
+   * @param {Object} content - partial object that will be used to patch the current content object
+   * @param {string} locale - locale key that the edits pertain to (Optional)
+   * @param {boolean} setModified - flag that decides whether this change will signal that state has been modified (Default: true)
+   * @returns
+   */
+  const setBlockContent = (content = null, locale = localeDefault, setModified = true) => {
+    if (setModified && !modified) setModified(true);
+    // don't update state if no content given
+    if (!content || Object.keys(content).length < 1) return;
+    // if block's content is not locale-dependent, set blockState.content
+    if (blockHasNoLocaleContent) {
+      setBlockState({...blockState, content: {...blockState.content, ...content}});
+    }
+    // else, edit the correct contentByLocale object
+    else {
+      const existingContent = blockState.contentByLocale[locale] || {id, locale, content: {}};
+      setBlockState({...blockState, contentByLocale: {...blockState.contentByLocale, [locale]: {...existingContent, content: {...existingContent.content, ...content}}}});
+    }
+  };
 
   const buildPayload = () => {
     const {settings} = blockState;
@@ -97,50 +121,22 @@ function Block({id, modified, callbacks, inline}) {
           inline ? setInlineId(null) : setOpened(false);
         }
       }
-      else {
-        // todo1.0 toast error
-      }
+      else { /* todo1.0 toast error*/ }
       setLoading(false);
     });
   };
 
-  /* CHANGE HANDLERS */
-
-  const upsertLocaleContent = (content, locale) => {
-    const existingContent = blockState.contentByLocale[locale] || {id, locale, content: {}};
-    setBlockState({...blockState, contentByLocale: {...blockState.contentByLocale, [locale]: {...existingContent, content: {...existingContent.content, ...content}}}});
-  };
-
-  const onChangeText = (content, locale) => upsertLocaleContent(content, locale);
-  const onTextModify = () => !modified && setModified(true);
-
-  const onChangeCode = (logic, locale) => {
-    if (!modified) setModified(true);
-    hasNoLocaleContent(block.type)
-      ? setBlockState({...blockState, content: {...blockState.content, logic}})
-      : upsertLocaleContent({logic}, locale);
-  };
-
-  const onChangeSimple = (simple, logic, locale) => {
-    if (!modified) setModified(true);
-    hasNoLocaleContent(block.type)
-      ? setBlockState({...blockState, content: {...blockState.content, simple, logic}})
-      : upsertLocaleContent({simple, logic}, locale);
-  };
+  const currentMode =
+    blockState.type && blockContent[BLOCK_FIELDS.SIMPLE_ENABLED]
+      ? MODES.UI
+      : MODES.CODE;
 
   const onChangeMode = (mode, locale) => {
-    if (!modified) setModified(true);
     const newMode = {[BLOCK_FIELDS.SIMPLE_ENABLED]: mode === MODES.UI};
-    hasNoLocaleContent(block.type)
-      ? setBlockState({...blockState, content: {...blockState.content, ...newMode}})
-      : upsertLocaleContent(newMode, locale);
+    setBlockContent(newMode, locale, true);
   };
 
-  const onChangeAPI = api => {
-    setBlockState({...blockState, content: {...blockState.content, api}});
-  };
-
-  const onChangeSettings = settings => {
+  const setBlockSettings = settings => {
     if (!modified) setModified(true);
     // Unlike other settings, type and shared are top-level, not stored in the settings object
     settings.type || settings.shared !== undefined
@@ -148,117 +144,18 @@ function Block({id, modified, callbacks, inline}) {
       : setBlockState({...blockState, settings: {...blockState.settings, ...settings}});
   };
 
-  /**
-   * A number of components embedded in this block need access to content here, either the ever-changing
-   * blockState, or the various callbacks that change it. It is recommended here https://reactjs.org/docs/composition-vs-inheritance.html
-   * to achieve this by passing these components down as props for deep embedding.
-   */
-
-  /* EDITORS TO PASS DOWN */
-
-  /**
-   * Gen/Viz/Selector editors should not re-render on every single change, as their javascript is often broken mid-keystroke.
-   * Therefore, render these blocks using the static props version, which only updates on Save. Normal stat-likes
-   * can update on keystroke, so the user can watch the prose change as they type.
-   */
-  // const usePropBlock = [BLOCK_TYPES.GENERATOR, BLOCK_TYPES.VIZ, BLOCK_TYPES.SELECTOR].includes(block.type);
-
-
-  const blockPreview = <BlockPreview
-    id={id}
-    key="block-preview"
-    active={true}
-    block={block}
-    blockState={blockState}
-    locale={localeDefault}
-    variables={variables}
-    allowed={true}
-    debug={true}
-  />;
-
-  const currentMode =
-    blockState.type  // The editor may not yet be open, don't try to drill down until the state is cloned.
-      ? (hasNoLocaleContent(blockState.type)
-        ? blockState.content
-        : blockState.contentByLocale[localeDefault].content)[BLOCK_FIELDS.SIMPLE_ENABLED]
-        ? MODES.UI
-        : MODES.CODE
-      : MODES.CODE;
-
-  const simpleState = hasNoLocaleContent(blockState.type)
-    ? blockState.content?.simple
-    : blockState.contentByLocale?.[localeDefault]?.content?.simple;
-
-  const apiInput = <ApiInput
-    key="api-input"
-    defaultValue={block.content.api}
-    onChange={onChangeAPI}
-    onEnterPress={() => onSave(true)}
-    variables={variables}
-  />;
-
-  /** Editor for modifying a block's JS logic directly */
-  // The codeEditor is the only editor that changes *based on another editor, i.e, the simpleUI.
-  // Therefore it must be controlled, and its state must live here in Block.jsx
-  const codeEditor = <AceWrapper
-    style={{flex: 1}}
-    key="code-editor"
-    onChange={logic => onChangeCode(logic, localeDefault)}
-    value={blockState.type
-      ? hasNoLocaleContent(blockState.type)
-        ? blockState.content.logic
-        : blockState.contentByLocale[localeDefault].content.logic
-      : ""
-    }
-  />;
-
-  const executeButton = <Button style={{minHeight: 40}} onClick={() => onSave(true)}>Save &amp; Execute</Button>;
-
-  const isStatlike = ![BLOCK_TYPES.GENERATOR, BLOCK_TYPES.VIZ, BLOCK_TYPES.SELECTOR].includes(block.type);
-
-  /** This is the (non-code) GUI editor for blocks that is meant for most simple cases.
-   * If the block is a stat-like type, then use a rich text editor to render its forms.
-   * Otherwise, a block type will need to have its own tailored Simple UI logic.
-   */
-  const uiEditor = isStatlike
-    ? <RichTextEditor
-      locale={localeDefault}
-      key="text-editor"
-      defaultContent={block.contentByLocale[localeDefault].content.simple || {}}
-      fields={BLOCK_MAP[block.type]}
-      formatters={formatters}
-      variables={variables}
-      onChange={onChangeText}
-      onTextModify={onTextModify}
-    />
-    : <SimpleUI
-      type={block.type}
-      locale={localeDefault}
-      simpleState={simpleState}
-      onChange={onChangeSimple}
-      executeButton={executeButton}
-    />;
-
-  const components = {blockPreview, apiInput, codeEditor, blockSettings, executeButton, uiEditor};
-
-  const blockSettings = <BlockSettings
-    id={id}
-    onChange={onChangeSettings}
-  />;
-
-  const blockOutputPanel = <BlockOutputPanel
-    id={id}
-    components={components}
-    mode={currentMode}
-  />;
-
-  const variableList = <VariableList id={id} setInlineId={setInlineId}/>;
-
-  const panels = {blockSettings, blockOutputPanel, variableList};
-
   return (
     <React.Fragment>
-      <BlockEditor key="be" id={id} panels={panels}/>
+      <BlockEditor
+        blockContent={blockContent}
+        currentMode={currentMode}
+        key="be"
+        id={id}
+        onSave={onSave}
+        setBlockContent={setBlockContent}
+        setBlockSettings={setBlockSettings}
+        variables={variables}
+      />
       <Group position="right">
         <Tooltip
           label={`${currentMode === MODES.UI ? "Show" : "Hide"} Code`}
